+++ ui/src/app/modules/login/login.module.ts
@@ -1,0 +1,21 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+
+import { LoginComponent } from './login.component';
+import { LoginGuard } from './login.guard';
+
+@NgModule({
+  declarations: [
+    LoginComponent,
+  ],
+  imports: [
+    CommonModule,
+    FormsModule,
+    ReactiveFormsModule,
+  ],
+  providers: [
+    LoginGuard,
+  ],
+})
+export class LoginModule { }