+++ ui/src/app/modules/setup-wizard/setup-wizard-routing.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+
+import { SetupWizardComponent } from './setup-wizard.component';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: SetupWizardComponent,
+  },
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class SetupWizardRoutingModule { }